backend/requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.0
astunparse==1.6.3
bcrypt==4.1.3
beautifulsoup4==4.14.2
black==25.9.0
boto3==1.40.50
botocore==1.40.50
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.2
curl_cffi==0.13.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.110.1
flake8==7.3.0
flatbuffers==25.9.23
frozendict==2.4.6
gast==0.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
h5py==3.15.0
httplib2==0.31.0
idna==3.10
iniconfig==2.1.0
isort==6.1.0
jmespath==1.0.1
joblib==1.5.2
jq==1.10.0
keras==3.11.3
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
ml_dtypes==0.5.3
motor==3.3.1
multitasking==0.0.12
mypy==1.18.2
mypy_extensions==1.1.0
namex==0.1.0
nltk==3.9.2
numpy==2.3.3
oauthlib==3.3.1
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathspec==0.12.1
patsy==1.0.1
peewee==3.18.2
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.5.0
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.1.10
pytz==2025.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rsa==4.9.1
s3transfer==0.14.0
s5cmd==0.2.0
scikit-learn==1.7.2
scipy==1.16.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.37.2
statsmodels==0.14.5
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
textblob==0.19.0
threadpoolctl==3.6.0
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.25.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
yfinance==0.2.66
